requirements.txt
requests
python-dotenv
com-digitalruiz-my-logger
